refactor(chronotype): 共有型と defaults を features/chronotype 配下に集約
lib/types/chronotype.ts と lib/chronotype-defaults.ts は consumer が全て

chronotype feature 内のみだった (+ calendar の grid type 1 箇所のみ barrel 経由化)。

src/lib から chronotype 専用残債を除去し、feature 内で type / defaults を

self-contained に管理する形に整理。

- git mv で 2 ファイル移動 (履歴保持)

  - lib/types/chronotype.ts → features/chronotype/types/chronotype.ts

  - lib/chronotype-defaults.ts → features/chronotype/lib/defaults.ts

- chronotype 内部 13 ファイルの import を相対 path に切替

- features/chronotype/index.ts barrel に ProductivityZone 追加

- features/calendar/types/grid.types.ts を barrel 経由 import に変更

仕様 / UI / DB / store 変更なし。boundary 違反なし (むしろ改善)。
