+++ apps/product/src/features/chronotype/lib/defaults.ts
@@ -1,4 +1,4 @@
-import type { ChronotypeSettings } from '@/lib/types/chronotype';
+import type { ChronotypeSettings } from '../types/chronotype';
 
 /** クロノタイプ設定のデフォルト値（null = 未設定） */
 export const DEFAULT_CHRONOTYPE_SETTINGS: ChronotypeSettings | null = null;
