+++ apps/product/src/features/chronotype/lib/quiz.ts
@@ -1,4 +1,4 @@
-import type { PresetChronotypeType } from '@/lib/types/chronotype';
+import type { PresetChronotypeType } from '../types/chronotype';
 
 /** 各選択肢のスコア配分 */
 export interface QuizOptionScores {
